requirements.txt
cryptography==1.5
jsonschema==2.5.1
lxml==3.6.4
numpy==1.11.3
pandas==0.21.1
pyasn1==0.1.9
pyarrow==0.7.1
Cython==0.25.2
pyOpenSSL==16.2.0
certifi==2017.1.23
channels==1.1.2
Django==1.11
ed25519ll==0.6
protobuf==3.2.0
keyring==10.2
mock==2.0.0
ndg_httpsclient==0.4.2
parameters==0.2.1
pytest==3.0.6
redis==2.10.5
requests_mock==1.3.0
djangorestframework==3.5.4
simplejson==3.10.0
wincertstore==0.2
pyxdg==0.25
xlrd==1.0.0
service_identity==16.0.0
psycopg2==2.7.1
tweepy==3.5.0
GitPython==2.1.3
django-user-accounts==2.0.3
django-polymorphic==1.3
django-analytical==2.2.2
nltk==3.2.4
twython==3.6.0
sendgrid-django==4.0.4
git+https://github.com/CJWorkbench/oauth2client.git#egg=oauth2client
httplib2==0.10.3
google-api-python-client==1.6.4
jsonpickle==0.9.5
